+++ flask_unchained/bundles/admin/extensions/__init__.py
@@ -1,0 +1,9 @@
+from .admin import Admin
+
+
+admin = Admin()
+
+
+EXTENSIONS = {
+    'admin': admin,
+}