+++ flask_unchained/bundles/admin/hooks/__init__.py
@@ -1,0 +1,3 @@
+class Store:
+    def __init__(self):
+        pass