+++ flask_unchained/bundles/admin/routes.py
@@ -1,0 +1,11 @@
+from flask_unchained import controller, rule
+
+from .views import AdminSecurityController
+
+
+routes = [
+    controller('/admin', AdminSecurityController, rules=[
+        rule('/login', 'login', endpoint='admin.login'),
+        rule('/logout', 'logout', endpoint='admin.logout'),
+    ]),
+]