+++ flask_unchained/bundles/admin/views/__init__.py
@@ -1,0 +1,2 @@
+from .admin_security_controller import AdminSecurityController
+from .dashboard import AdminDashboardView